Extract common SortableModel
